ThreadPoolManager tweak:
From now Runnable tasks don't have to be wrapped into try-catch blocks, if it was used only for logging purpose, TPM do it automatically ...and the ScheduledThreadPoolExecutors "won't eat" Exceptions anymore. You will see a "nice" log :)

BTW, why the f*ckin' hell we need so many thread pools? There is no reason to have it...

About Exceptions: what I saw in some changesets in the last days...

If you write this:
_log.warn(e)
It will call e.toString() method, what gives back ONLY the name of the exception, like "NullPointerException", so don't use it. If you want to make a log, make a descriptive one, or nothing at all.

So imho use this, if you want to print the stack trace:
_log.warn("a message whatever you want here", e);
It will print at the message, and the stack trace too...

... and DO NOT USE this (except if logging doesn't work at point - like shutdown):
e.printStackTrace();
If the exception "generated" inside a big loop this will flood the log, although "_log" methods won't print the stackTrace of that specific exception after x appearance.

p.s.: test this changes with care... I'm using my an own-tpm, so I can't test this one :)

git-svn-id: http://svn.l2jfree.com/l2jfree/trunk@3872 6ff00ba9-25cd-4355-987e-099c62bc2628